working example with randomAgent
