+++ reinforcement_learning/demos/cipher_puzzle/data/README.txt
@@ -1,0 +1,2 @@
+famous_quotes.txt - from https://github.com/alvations/Quotables
+word_list.txt - from nltk words corpus
