--- reinforcement_learning/main.py
@@ -1,6 +1,0 @@
-import click
-
-@click.command()
-def cli():
-    """Example script."""
-    click.echo('Hello World!')
